Add support for timing functions jump-start|end|both|none.
Jump-start is equivalent to start, with a discontinuity at t = 0.
Jump-end is equivalent to end, with a discontinuity at t=1.
Jump-both has discontinuities at t=0 and 1.
Jump-none is continuous at t=0 and t=1.

The patch also addresses some serialization issues with existing
step functions since the code needed to be updated in support of the
new step types.

https://drafts.csswg.org/css-easing-1/#step-easing-functions

https://www.chromestatus.com/feature/5730327525851136

Bug: 827560
Change-Id: I950ed09b2ca72c9e09a4f5e3f8d373cc5acbd07b