[LayoutNG] Hang preserved trailing spaces for `text-align: start`
This patch hangs preserved trailing spaces when `text-align`
is `start` (or `left` in LTR or `right` in RTL).

The previous tests did not cover `start` case because it was
thought not observable, but issue 966773 discovered that it
is ovservable, and tests figured out that all impls hang in
this case.

This patch matches LayoutNG to legacy. The test is tentative
until the CSS WG resolution.

One test was failing before this patch but was accidentally
passed. Now it really fails:

Bug: 966773
Change-Id: I324b0c83e7aa9e5f3233b7bc56b262ddc5ee1fe2